src/test/java/com/devweb/devweb/services/RecipeServiceImplTest.java
package com.devweb.devweb.services;

import com.devweb.devweb.domain.Recipe;
import com.devweb.devweb.repositories.RecipeRepository;
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.HashSet;
import java.util.Set;

import static org.junit.Assert.*;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

public class RecipeServiceImplTest {

    RecipeServiceImpl recipeService;
    @Mock
    RecipeRepository recipeRepository;


    @Before
    public void setUp() throws Exception {
        MockitoAnnotations.initMocks(this);

        recipeService = new RecipeServiceImpl(recipeRepository);
    }

    @Test
    public void getrecipes() {
        Recipe recipe = new Recipe();
        HashSet recipesData = new HashSet();
        recipesData.add(recipe);
        when(recipeService.getrecipes()).thenReturn(recipesData);

        Set<Recipe> recipes = recipeService.getrecipes();
        assertEquals(recipes.size(),1);
            verify(recipeRepository,times(1)).findAll();
    }
}